JS - changed the intent to receive an int from PeopleList.  changed the key in the intent to be personId.

+++ src/main/java/com/thoughtworks/carpods/UI/pod/EditPod.java
@@ -104,7 +104,7 @@ public class EditPod extends PodActivity {
     public void onActivityResult(int requestCode, int resultCode, Intent data) {
         if (resultCode == RESULT_OK) {
             if (requestCode == PICK_CONTACT_REQUEST) {
-                long personId = data.getLongExtra("personId", -1);
+                long personId = data.getIntExtra("personId", -1);
                 PeopleDataAccess personDataAccess = dataAccessFor.people(this);
                 Person newPodMember = personDataAccess.getPersonFromDatabaseWithId(personId);
                 addMemberToMemberList(newPodMember);
